backend/requirements.txt
# FastAPI 核心依赖
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# 数据验证
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# PDF 处理
PyMuPDF>=1.23.0

# 图像处理
Pillow>=10.3.0

# 音频处理
faster-whisper>=1.0.0

# PDF 生成
reportlab>=4.1.0

# HTTP 请求
requests>=2.31.0
httpx>=0.26.0

# 环境变量
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# 认证
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
loguru>=0.7.2

# 异步支持
aiofiles>=23.2.0

# OpenAI API (可选)
# openai==1.3.0

# 其他工具
numpy>=1.26.0
python-json-logger==2.0.7
